server-api/requirements.txt
websockets>=12.0
google-auth>=2.23.0
certifi>=2023.7.22
requests>=2.31.0
google-cloud-storage>=2.13.0
fastapi>=0.104.0
uvicorn>=0.23.2